project/requirements.txt
python-pptx==0.6.21
openai==1.3.0
python-docx==0.8.11
PyPDF2==3.0.1
langchain==0.0.350
matplotlib==3.7.1
python-dotenv==1.0.0
tqdm==4.65.0
numpy==1.26.4
scipy==1.15.2
pandas==2.2.0
markdown==3.4.3
markdown-it-py==2.2.0
markdown-it-charts==0.1.0
jinja2==3.1.2
pyyaml==6.0
beautifulsoup4==4.12.2
pillow==9.5.0 